rework task and generators
This should close #3

The tasks are now a generated folder `_tasks` instead of a standard
grunt plugin, you can then tweak or add new jekyll related tasks in
there.

To update, npm update the package and rerun the generator should do it.
The diffing of conflicts should give good hints.